+++ api/src/test/java/org/jboss/cdi/api/test/AnnotationLiteralTest.java
@@ -14,7 +14,8 @@
 package org.jboss.cdi.api.test;
 
 import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertTrue;
+import static org.testng.Assert.assertNotEquals;
+import static org.testng.Assert.assertNotNull;
 
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.enterprise.context.BeforeDestroyed;
@@ -67,21 +68,40 @@ public class AnnotationLiteralTest {
         new FooLiteral(null).hashCode();
     }
 
-    @SuppressWarnings("serial")
+    @Test
+    public void toStringShouldWork() {
+        assertNotNull(new FooLiteral("foo").toString());
+    }
+
+    @Test
+    public void annotationTypeShouldWork() {
+        assertEquals(Foo.class, new FooLiteral("foo").annotationType());
+    }
+
+    @Test
+    public void equalsShouldWork() {
+        FooLiteral foo = new FooLiteral("foo");
+        FooLiteral foo2 = new FooLiteral("foo");
+        assertEquals(foo, foo2);
+    }
+
+    @Test
+    public void hashCodeShouldWork() {
+        assertNotEquals(new FooLiteral("foo").hashCode(), 0);
+    }
+
     @Test
     public void testDefaultLiteral() {
         assertEquals(new AnnotationLiteral<Default>() {
         }, Default.Literal.INSTANCE);
     }
 
-    @SuppressWarnings("serial")
     @Test
     public void testAnyLiteral() {
         assertEquals(new AnnotationLiteral<Any>() {
         }, Any.Literal.INSTANCE);
     }
 
-    @SuppressWarnings("serial")
     @Test
     public void testNonbindingLiteral() {
         assertEquals(new AnnotationLiteral<Nonbinding>() {
@@ -90,11 +110,10 @@ public class AnnotationLiteralTest {
 
     @Test
     public void testTypedLiteral() {
-        assertTrue(Typed.Literal.INSTANCE.value().length == 0);
-        assertTrue(Typed.Literal.of(new Class[] { String.class }).value()[0] == String.class);
+        assertEquals(Typed.Literal.INSTANCE.value().length, 0);
+        assertEquals(Typed.Literal.of(new Class[] { String.class }).value()[0], String.class);
     }
 
-    @SuppressWarnings("serial")
     @Test
     public void testAlternativeLiteral() {
         assertEquals(new AnnotationLiteral<Alternative>() {
@@ -107,14 +126,12 @@ public class AnnotationLiteralTest {
         assertEquals(NamedLiteral.of("foo").value(), "foo");
     }
 
-    @SuppressWarnings("serial")
     @Test
     public void testQualifierLiteral() {
         assertEquals(new AnnotationLiteral<Qualifier>() {
         }, QualifierLiteral.INSTANCE);
     }
 
-    @SuppressWarnings("serial")
     @Test
     public void testSingletonLiteral() {
         assertEquals(new AnnotationLiteral<Singleton>() {
@@ -144,63 +161,54 @@ public class AnnotationLiteralTest {
         assertEquals(BeforeDestroyed.Literal.APPLICATION.value(), ApplicationScoped.class);
     }
 
-    @SuppressWarnings("serial")
     @Test
     public void testApplicationScopedLiteral() {
         assertEquals(new AnnotationLiteral<ApplicationScoped>() {
         }, ApplicationScoped.Literal.INSTANCE);
     }
 
-    @SuppressWarnings("serial")
     @Test
     public void testRequestScopedLiteral() {
         assertEquals(new AnnotationLiteral<RequestScoped>() {
         }, RequestScoped.Literal.INSTANCE);
     }
 
-    @SuppressWarnings("serial")
     @Test
     public void testSessionScopedLiteral() {
         assertEquals(new AnnotationLiteral<SessionScoped>() {
         }, SessionScoped.Literal.INSTANCE);
     }
 
-    @SuppressWarnings("serial")
     @Test
     public void testConversationScopedLiteral() {
         assertEquals(new AnnotationLiteral<ConversationScoped>() {
         }, ConversationScoped.Literal.INSTANCE);
     }
 
-    @SuppressWarnings("serial")
     @Test
     public void testDependentLiteral() {
         assertEquals(new AnnotationLiteral<Dependent>() {
         }, Dependent.Literal.INSTANCE);
     }
 
-    @SuppressWarnings("serial")
     @Test
     public void testVetoedLiteral() {
         assertEquals(new AnnotationLiteral<Vetoed>() {
         }, Vetoed.Literal.INSTANCE);
     }
 
-    @SuppressWarnings("serial")
     @Test
     public void testInjectLiteral() {
         assertEquals(new AnnotationLiteral<Inject>() {
         }, InjectLiteral.INSTANCE);
     }
 
-    @SuppressWarnings("serial")
     @Test
     public void testSpecializesLiteral() {
         assertEquals(new AnnotationLiteral<Specializes>() {
         }, Specializes.Literal.INSTANCE);
     }
 
-    @SuppressWarnings("serial")
     @Test
     public void testTransientReferenceLiteral() {
         assertEquals(new AnnotationLiteral<TransientReference>() {
