--- api/src/test/resources/META-INF/services/fake
@@ -1,1 +1,0 @@
-org.jboss.cdi.api.test.privileged.FakeCDIProvider
