--- api/src/test/resources/java.policy
@@ -1,75 +1,0 @@
-/*
- * Copyright (c) 2018 Red Hat, Inc. and others
- * 
- * This program and the accompanying materials are made available under the
- * Apache Software License 2.0 which is available at:
- * https://www.apache.org/licenses/LICENSE-2.0.
- * 
- * SPDX-License-Identifier: Apache-2.0
- */
-// Standard extensions get all permissions by default
-
-grant codeBase "file:${{java.ext.dirs}}/*" {
-        permission java.security.AllPermission;
-};
-
-// Grant all permissions to jar in local repo to have no issue with maven surefire and testng
-
-grant codeBase "file:${settings.localRepository}/-" {
-        permission java.security.AllPermission;
-};
-
-// Grant all permission to CDI API classes. As test classes will have only standard permission, privileged bloc will be useful
-
-grant codeBase "file:${project.build.outputDirectory}/-" {
-        permission java.security.AllPermission;
-};
-
-grant codeBase "file:${project.build.testOutputDirectory}/-" {
-        permission java.util.PropertyPermission "serviceDir", "read";
-        permission java.io.FilePermission "${project.build.testOutputDirectory}/META-INF/services/jakarta.enterprise.inject.spi.CDIProvider", "write";
-};
-
-
-// default permissions granted to all domains
-
-grant {
-        // Allows any thread to stop itself using the java.lang.Thread.stop()
-        // method that takes no argument.
-        // Note that this permission is granted by default only to remain
-        // backwards compatible.
-        // It is strongly recommended that you either remove this permission
-        // from this policy file or further restrict it to code sources
-        // that you specify, because Thread.stop() is potentially unsafe.
-        // See the API specification of java.lang.Thread.stop() for more
-        // information.
-        permission java.lang.RuntimePermission "stopThread";
-
-        // allows anyone to listen on dynamic ports
-        permission java.net.SocketPermission "localhost:0", "listen";
-
-        // "standard" properties that can be read by anyone
-
-        permission java.util.PropertyPermission "java.version", "read";
-        permission java.util.PropertyPermission "java.vendor", "read";
-        permission java.util.PropertyPermission "java.vendor.url", "read";
-        permission java.util.PropertyPermission "java.class.version", "read";
-        permission java.util.PropertyPermission "os.name", "read";
-        permission java.util.PropertyPermission "os.version", "read";
-        permission java.util.PropertyPermission "os.arch", "read";
-        permission java.util.PropertyPermission "file.separator", "read";
-        permission java.util.PropertyPermission "path.separator", "read";
-        permission java.util.PropertyPermission "line.separator", "read";
-
-        permission java.util.PropertyPermission "java.specification.version", "read";
-        permission java.util.PropertyPermission "java.specification.vendor", "read";
-        permission java.util.PropertyPermission "java.specification.name", "read";
-
-        permission java.util.PropertyPermission "java.vm.specification.version", "read";
-        permission java.util.PropertyPermission "java.vm.specification.vendor", "read";
-        permission java.util.PropertyPermission "java.vm.specification.name", "read";
-        permission java.util.PropertyPermission "java.vm.version", "read";
-        permission java.util.PropertyPermission "java.vm.vendor", "read";
-        permission java.util.PropertyPermission "java.vm.name", "read";
-};
-
